Adapted code for Tel Aviv model to better integrate external trips.
git-svn-id: file:///srv/local/data/cleanRepo/trunk@23777 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/telaviv/src/main/java/playground/telaviv/population/Emme2ExternalTrip.java
@@ -27,6 +27,6 @@ public class Emme2ExternalTrip {
 	
 	int originNodeId;
 	int destinationNodeId;
-	int numOfTrips;
+	double numOfTrips;
 	
 }
